Fix BLSubSample attribute type to long
Primary key for BLSubSample set to type long
Update DataCollectionInfo mapper and tests

+++ uk.ac.diamond.ispyb.api/src/uk/ac/diamond/ispyb/api/IspybPlateApi.java
@@ -27,7 +27,7 @@ public interface IspybPlateApi extends Closeable {
 
 	void insertContainerError(String barcode, String error, int severity, String stackTrace) throws SQLException;
 
-	public List<DataCollectionInfo> retrieveDataCollectionInfosForSubsample(int id);
+	public List<DataCollectionInfo> retrieveDataCollectionInfosForSubsample(long id);
 	
 	void clearContainerError(String barcode) throws SQLException;
